refactor(agent): consolidate agent service methods and update imports for improved structure
