+++ web-app/src/app/(app)/agents/[agentId]/jobs/default.tsx
@@ -1,7 +1,7 @@
 import { notFound } from "next/navigation";
 
 import { getSessionOrRedirect } from "@/lib/auth/utils";
-import { getAgentById, getMyJobsByAgentId } from "@/lib/services";
+import { AgentService, getMyJobsByAgentId } from "@/lib/services";
 
 import JobsTable from "./components/jobs-table";
 
@@ -13,7 +13,7 @@ export default async function JobsPage({ params }: JobsPageProps) {
   const session = await getSessionOrRedirect();
 
   const { agentId } = await params;
-  const agent = await getAgentById(agentId);
+  const agent = await AgentService.getInstance().getAgentById(agentId);
   if (!agent) {
     console.warn("agent not found in job layout");
     return notFound();
