+++ web-app/src/app/(app)/agents/page.tsx
@@ -3,11 +3,7 @@ import { getTranslations } from "next-intl/server";
 
 import { AgentsNotAvailable } from "@/components/agents";
 import { getSessionOrRedirect } from "@/lib/auth/utils";
-import {
-  getAvailableAgentsWithCreditsPrice,
-  getFavoriteAgents,
-  TagService,
-} from "@/lib/services";
+import { AgentService, TagService } from "@/lib/services";
 import { Tag } from "@/prisma/generated/client";
 
 import FilterSection from "./components/filter-section";
@@ -25,7 +21,9 @@ export async function generateMetadata(): Promise<Metadata> {
 export default async function GalleryPage() {
   await getSessionOrRedirect();
 
-  const agentsWithPrice = await getAvailableAgentsWithCreditsPrice();
+  const agentService = AgentService.getInstance();
+  const agentsWithPrice =
+    await agentService.getAvailableAgentsWithCreditsPrice();
 
   if (!agentsWithPrice.length) {
     return <AgentsNotAvailable />;
@@ -34,7 +32,7 @@ export default async function GalleryPage() {
   const tags: Tag[] = await TagService.getInstance().getActiveTags();
   const tagNames = tags.map((tag) => tag.name);
 
-  const favoriteAgents = await getFavoriteAgents();
+  const favoriteAgents = await agentService.getFavoriteAgents();
 
   return (
     <div className="w-full">
