+++ web-app/src/lib/services/agent.service.ts
@@ -4,12 +4,15 @@ import * as Sentry from "@sentry/nextjs";
 
 import { getEnvPublicConfig } from "@/config/env.public";
 import { getEnvSecrets } from "@/config/env.secrets";
+import { getSession, getSessionOrThrow } from "@/lib/auth/utils";
 import {
   agentInclude,
   agentListInclude,
   AgentListWithAgents,
   agentOrderBy,
+  AgentWithFixedPricing,
   AgentWithJobs,
+  AgentWithOrganizations,
   AgentWithRelations,
 } from "@/lib/db/types";
 import {
@@ -23,13 +26,252 @@ import {
   AgentList,
   AgentListType,
   AgentStatus,
+  CreditCost,
 } from "@/prisma/generated/client";
 
 import { BaseService } from "./base.service";
+import { getAgentCreditsPrice } from "./credit";
+import { CreditCostService } from "./creditCost.service";
+import { MemberService } from "./member.service";
+
+export interface AgentWithCreditPrice {
+  agent: AgentWithRelations;
+  creditsPrice: Awaited<ReturnType<typeof getAgentCreditsPrice>>;
+}
 
 export class AgentService extends BaseService<AgentService> {
   private static thresholdDays =
     getEnvPublicConfig().NEXT_PUBLIC_AGENT_NEW_THRESHOLD_DAYS;
+
+  /**
+   * Utility: Checks if a user can access an agent based on organization membership and agent visibility.
+   *
+   * @param agent - Agent with organization data.
+   * @param userOrganizationIds - Organization IDs the user is a member of.
+   * @returns True if the user can access the agent, false otherwise.
+   */
+  private canUserAccessAgent(
+    agent: AgentWithOrganizations,
+    userOrganizationIds: string[],
+  ): boolean {
+    if (!agent.isShown) return false;
+    if (agent.organizations.length === 0) return true;
+    if (userOrganizationIds.length === 0) return false;
+    return agent.organizations.some((agentOrg) =>
+      userOrganizationIds.includes(agentOrg.id),
+    );
+  }
+
+  /**
+   * Utility: Checks if an agent's fixed pricing units are all valid according to the provided credit costs.
+   *
+   * @param agent - Agent with fixed pricing information.
+   * @param creditCosts - Array of valid credit cost objects.
+   * @returns True if all pricing units are valid or if there are no amounts, false otherwise.
+   */
+  private hasValidPricing(
+    agent: AgentWithFixedPricing,
+    creditCosts: CreditCost[],
+  ): boolean {
+    const units = creditCosts.map(({ unit }) => unit);
+    const amounts = agent.pricing.fixedPricing?.amounts?.map((amount) => ({
+      unit: amount.unit,
+      amount: Number(amount.amount),
+    }));
+    if (!amounts) {
+      return true;
+    }
+    return amounts.every(({ unit }) => units.includes(unit));
+  }
+
+  /**
+   * Utility: Determines if an agent is available to the user based on access permissions and pricing validity.
+   *
+   * @param agent - Agent with relations including organization and pricing data.
+   * @param organizationIds - Organization IDs the user is a member of.
+   * @param creditCosts - Valid credit cost objects for pricing validation.
+   * @returns True if the agent is available to the user, false otherwise.
+   */
+  private isAgentAvailable(
+    agent: AgentWithRelations,
+    organizationIds: string[],
+    creditCosts: CreditCost[],
+  ): boolean {
+    return (
+      this.canUserAccessAgent(agent, organizationIds) &&
+      this.hasValidPricing(agent, creditCosts)
+    );
+  }
+
+  /**
+   * Retrieves the current session's organization IDs and all credit costs for agent access checks.
+   *
+   * @param tx - Optional Prisma transaction client for DB operations.
+   * @returns Object with userOrganizationIds and creditCosts.
+   */
+  private async getAgentAccessContext(): Promise<{
+    userOrganizationIds: string[];
+    creditCosts: CreditCost[];
+  }> {
+    const session = await getSession();
+    const creditCosts = await CreditCostService.getInstance(
+      this.client,
+    ).getCreditCosts();
+    const userOrganizationIds =
+      session?.user.id && session.user.id !== ""
+        ? await MemberService.getInstance(
+            this.client,
+          ).getMembersOrganizationIdsByUserId(session.user.id)
+        : [];
+    return { userOrganizationIds, creditCosts };
+  }
+
+  // Service
+
+  async getAvailableAgentById(
+    agentId: string,
+  ): Promise<AgentWithRelations | null> {
+    const agent = await this.getShownAgentWithRelationById(
+      agentId,
+      AgentStatus.ONLINE,
+    );
+    if (!agent) return null;
+    const { userOrganizationIds, creditCosts } =
+      await this.getAgentAccessContext();
+    if (!this.isAgentAvailable(agent, userOrganizationIds, creditCosts))
+      return null;
+    return agent;
+  }
+
+  /**
+   * Retrieves an agent by ID with all relations, without access control.
+   *
+   * @param agentId - Unique agent identifier.
+   * @param tx - Optional Prisma transaction client.
+   * @returns The agent with all relations, or null if not found.
+   */
+  async getAgentById(agentId: string): Promise<AgentWithRelations | null> {
+    return await this.getAgentWithRelationsById(agentId);
+  }
+
+  /**
+   * Retrieves all online agents available to the current user with valid pricing.
+   *
+   * @param tx - Optional Prisma transaction client.
+   * @returns Array of available agents with valid pricing.
+   */
+  async getAvailableAgents(): Promise<AgentWithRelations[]> {
+    const { userOrganizationIds, creditCosts } =
+      await this.getAgentAccessContext();
+    const onlineAgents = await this.getShownAgentsWithRelationsByStatus(
+      AgentStatus.ONLINE,
+    );
+    return onlineAgents.filter((agent) =>
+      this.isAgentAvailable(agent, userOrganizationIds, creditCosts),
+    );
+  }
+
+  /**
+   * Retrieves all online agents available to the user, each with its calculated credit price.
+   *
+   * - Excludes agents for which credit price calculation fails.
+   *
+   * @param tx - Optional Prisma transaction client.
+   * @returns Array of agents with their calculated credit prices.
+   */
+  async getAvailableAgentsWithCreditsPrice(): Promise<AgentWithCreditPrice[]> {
+    const agents = await this.getAvailableAgents();
+    const results = await Promise.allSettled(
+      agents.map(async (agent) => {
+        const creditsPrice = await getAgentCreditsPrice(agent);
+        return { agent, creditsPrice };
+      }),
+    );
+    return results
+      .filter(
+        (result): result is PromiseFulfilledResult<AgentWithCreditPrice> =>
+          result.status === "fulfilled",
+      )
+      .map((result) => result.value);
+  }
+
+  /**
+   * Retrieves all agents hired by the current user, ordered by the most recent job activity (newest first).
+   *
+   * - Requires an active user session.
+   * - Agents without jobs are placed at the end of the list.
+   *
+   * @returns Array of hired agents with their jobs, sorted by recent activity.
+   * @throws If no active session is found.
+   */
+  async getHiredAgentsOrderedByLatestJob(): Promise<AgentWithJobs[]> {
+    const session = await getSessionOrThrow();
+    const userId = session.user.id;
+    const activeOrganizationId = session.session.activeOrganizationId;
+    const hiredAgentsWithJobs =
+      await this.getHiredAgentsWithJobsByUserIdAndOrganization(
+        userId,
+        activeOrganizationId,
+      );
+    return hiredAgentsWithJobs.sort((a, b) => {
+      const aLatestJob = a.jobs[0];
+      const bLatestJob = b.jobs[0];
+      if (!aLatestJob) return 1;
+      if (!bLatestJob) return -1;
+      return bLatestJob.startedAt.getTime() - aLatestJob.startedAt.getTime();
+    });
+  }
+
+  /**
+   * Retrieves the input schema definition for a specific agent, used to validate job inputs.
+   *
+   * - Throws an error if the agent or schema cannot be found.
+   *
+   * @param agentId - Unique agent identifier.
+   * @returns The agent's input schema definition.
+   * @throws If the agent is not found or if the schema cannot be fetched.
+   */
+  async getAgentInputSchema(agentId: string): Promise<JobInputsDataSchemaType> {
+    const agent = await this.getAgentWithRelationsById(agentId);
+    if (!agent) {
+      throw new Error(`Agent with ID ${agentId} not found`);
+    }
+    const inputSchemaResult = await this.fetchAgentInputSchema(agent);
+    if (!inputSchemaResult.ok) {
+      throw new Error(inputSchemaResult.error);
+    }
+    return inputSchemaResult.data;
+  }
+
+  async getFavoriteAgents(): Promise<AgentWithRelations[]> {
+    return await this.getAgentsByListType(AgentListType.FAVORITE);
+  }
+
+  private async getAgentsByListType(
+    type: AgentListType,
+  ): Promise<AgentWithRelations[]> {
+    const session = await getSessionOrThrow();
+    const existingList = await this.getAgentListByUserIdAndType(
+      session.user.id,
+      type,
+    );
+    if (existingList) {
+      const { userOrganizationIds, creditCosts } =
+        await this.getAgentAccessContext();
+      return existingList.agents
+        .map(AgentService.mapAgentWithIsNew)
+        .filter((agent) =>
+          this.isAgentAvailable(agent, userOrganizationIds, creditCosts),
+        );
+    }
+    const list = await this.createAgentListByUserIdAndType(
+      session.user.id,
+      type,
+    );
+    return list.agents.map(AgentService.mapAgentWithIsNew);
+  }
+
+  // Repo
 
   async getAgentsWithRelations(): Promise<AgentWithRelations[]> {
     const agents = await this.client.agent.findMany({
