--- web-app/src/lib/services/agent/index.ts
@@ -1,1 +1,0 @@
-export * from "./service";
